requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
sentence-transformers==2.6.1
faiss-cpu==1.7.4
numpy==1.26.4
transformers>=4.37.0
peft>=0.9.0
accelerate>=0.26.0
torch>=2.1.0
httpx>=0.27.0
python-dotenv>=1.0.1
jinja2>=3.1.3
rich>=13.7.0
openai>=1.12.0
langchain>=0.2.7
langchain-community>=0.2.7
langchain-openai>=0.1.7
psycopg2-binary>=2.9.9
pgvector>=0.3.2
qianfan>=0.3.8
pymysql>=1.1.0
pymongo>=4.7.3
